Update codebase for Prisma integration and unit tests additions
Adapted various database models, routes, and controllers to interface with a Prisma database rather than the previous knex-based system. Moreover, added new unit tests files to test the functions associated with Prisma database models. Updated package.json and package-lock.json files with a new devDependency entry for sinon. The sinon is used as a the test double library in the unit tests.