Created Dockerfile, updated package.json, and added .dockerignore file
This commit introduces a Dockerfile for containerizing the application. It also updates the scripts and dependencies in the package.json. Lastly, the commit includes a .dockerignore file to exclude unnecessary files from the Docker build context. Changes also include use of additional babel packages and reflect in their respective lock files.